Add existing code
This is the raw code accumulated over the years in various locations and now
gathered in one place. Some of it fails to build, and some parts are missing
for various reasons. Before I spend time on cleaning all this up, I better get
it consolidated in one place even though it's not perfect. After all, the
perfect is the enemy of the good, as they saying goes ;-)